refactor: move Lotto class and LottoTest back to the root package.
