+++ src/main/kotlin/lotto/view/InputView.kt
@@ -1,7 +1,7 @@
 package lotto.view
 
 import camp.nextstep.edu.missionutils.Console
-import lotto.model.Lotto
+import lotto.Lotto
 
 object InputView {
     fun purchaseAmount(): Int {
